+++ scm-core/src/main/java/sonia/scm/store/ConfigurationStoreFactory.java
@@ -21,11 +21,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-    
+
 package sonia.scm.store;
 
-
-import sonia.scm.repository.Repository;
 
 /**
  * The ConfigurationStoreFactory can be used to create new or get existing {@link ConfigurationStore} objects.
@@ -51,7 +49,7 @@ import sonia.scm.repository.Repository;
  * </pre></code>
  *
  * @author Sebastian Sdorra
- * 
+ *
  * @apiviz.landmark
  * @apiviz.uses sonia.scm.store.ConfigurationStore
  */
@@ -72,67 +70,7 @@ public interface ConfigurationStoreFactory  {
    * @return Floating API to set the name and either specify a repository or directly build a global
    * {@link ConfigurationStore}.
    */
-  default <T> TypedFloatingConfigurationStoreParameters<T>.Builder withType(Class<T> type) {
-    return new TypedFloatingConfigurationStoreParameters<T>(this).new Builder(type);
-  }
-}
-
-final class TypedFloatingConfigurationStoreParameters<T> {
-
-  private final TypedStoreParametersImpl<T> parameters = new TypedStoreParametersImpl<>();
-  private final ConfigurationStoreFactory factory;
-
-  TypedFloatingConfigurationStoreParameters(ConfigurationStoreFactory factory) {
-    this.factory = factory;
-  }
-
-  public class Builder {
-
-    Builder(Class<T> type) {
-      parameters.setType(type);
-    }
-
-    /**
-     * Use this to set the name for the {@link ConfigurationStore}.
-     * @param name The name for the {@link ConfigurationStore}.
-     * @return Floating API to either specify a repository or directly build a global {@link ConfigurationStore}.
-     */
-    public OptionalRepositoryBuilder withName(String name) {
-      parameters.setName(name);
-      return new OptionalRepositoryBuilder();
-    }
-  }
-
-  public class OptionalRepositoryBuilder {
-
-    /**
-     * Use this to create or get a {@link ConfigurationStore} for a specific repository. This step is optional. If you
-     * want to have a global {@link ConfigurationStore}, omit this.
-     * @param repository The optional repository for the {@link ConfigurationStore}.
-     * @return Floating API to finish the call.
-     */
-    public OptionalRepositoryBuilder forRepository(Repository repository) {
-      parameters.setRepositoryId(repository.getId());
-      return this;
-    }
-
-    /**
-     * Use this to create or get a {@link ConfigurationStore} for a specific repository. This step is optional. If you
-     * want to have a global {@link ConfigurationStore}, omit this.
-     * @param repositoryId The id of the optional repository for the {@link ConfigurationStore}.
-     * @return Floating API to finish the call.
-     */
-    public OptionalRepositoryBuilder forRepository(String repositoryId) {
-      parameters.setRepositoryId(repositoryId);
-      return this;
-    }
-
-    /**
-     * Creates or gets the {@link ConfigurationStore} with the given name and (if specified) the given repository. If no
-     * repository is given, the {@link ConfigurationStore} will be global.
-     */
-    public ConfigurationStore<T> build(){
-      return factory.getStore(parameters);
-    }
+  default <T> TypedStoreParametersBuilder<T,ConfigurationStore<T>> withType(Class<T> type) {
+    return new TypedStoreParametersBuilder<>(type, this::getStore);
   }
 }